+++ lib/src/ast/property.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 import 'package:angular_ast/src/ast.dart';
 import 'package:angular_ast/src/token.dart';
+import 'package:angular_ast/src/visitor.dart';
 import 'package:source_span/source_span.dart';
 import 'package:quiver/core.dart';
 
@@ -49,6 +50,11 @@ abstract class PropertyAst implements TemplateAst {
 
   @override
   int get hashCode => hash4(expression, name, postfix, unit);
+
+  @override
+  /*=R*/ accept/*<R, C>*/(TemplateAstVisitor/*<R, C>*/ visitor, [C context]) {
+    return visitor.visitProperty(this, context);
+  }
 
   /// Bound expression; optional for backwards compatibility.
   ExpressionAst get expression;