+++ lib/src/ast/reference.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 import 'package:angular_ast/src/ast.dart';
 import 'package:angular_ast/src/token.dart';
+import 'package:angular_ast/src/visitor.dart';
 import 'package:source_span/source_span.dart';
 import 'package:quiver/core.dart';
 
@@ -41,6 +42,11 @@ abstract class ReferenceAst implements TemplateAst {
 
   @override
   int get hashCode => hash2(identifier, variable);
+
+  @override
+  /*=R*/ accept/*<R, C>*/(TemplateAstVisitor/*<R, C>*/ visitor, [C context]) {
+    return visitor.visitReference(this, context);
+  }
 
   /// What `exportAs` identifier to assign to [variable].
   ///