+++ lib/src/ast/sugar/banana.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 import 'package:angular_ast/src/ast.dart';
 import 'package:angular_ast/src/token.dart';
+import 'package:angular_ast/src/visitor.dart';
 import 'package:source_span/source_span.dart';
 import 'package:quiver/core.dart';
 
@@ -33,6 +34,11 @@ abstract class BananaAst implements TemplateAst {
     NgToken fieldToken,
     NgToken endFieldToken,
   ) = _ParsedBananaAst;
+
+  @override
+  /*=R*/ accept/*<R, C>*/(TemplateAstVisitor/*<R, C>*/ visitor, [C context]) {
+    throw new UnimplementedError();
+  }
 
   @override
   bool operator ==(Object o) {