+++ lib/src/visitors/identity.dart
@@ -1,0 +1,49 @@
+// Copyright (c) 2016, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:angular_ast/src/ast.dart';
+import 'package:angular_ast/src/visitor.dart';
+import 'package:meta/meta.dart';
+
+/// An [TemplateAstVisitor] that does nothing but return the AST node back.
+class IdentityTemplateAstVisitor<C>
+    implements TemplateAstVisitor<TemplateAst, C> {
+  @literal
+  const IdentityTemplateAstVisitor();
+
+  @override
+  TemplateAst visit(TemplateAst astNode, [_]) => astNode;
+
+  @override
+  TemplateAst visitAttribute(AttributeAst astNode, [_]) => astNode;
+
+  @override
+  TemplateAst visitComment(CommentAst astNode, [_]) => astNode;
+
+  @override
+  TemplateAst visitEmbeddedContent(EmbeddedContentAst astNode, [_]) => astNode;
+
+  @override
+  TemplateAst visitEmbeddedTemplate(EmbeddedTemplateAst astNode, [_]) {
+    return astNode;
+  }
+
+  @override
+  TemplateAst visitElement(ElementAst astNode, [_]) => astNode;
+
+  @override
+  TemplateAst visitEvent(EventAst astNode, [_]) => astNode;
+
+  @override
+  TemplateAst visitInterpolation(InterpolationAst astNode, [_]) => astNode;
+
+  @override
+  TemplateAst visitProperty(PropertyAst astNode, [_]) => astNode;
+
+  @override
+  TemplateAst visitReference(ReferenceAst astNode, [_]) => astNode;
+
+  @override
+  TemplateAst visitText(TextAst astNode, [_]) => astNode;
+}